we back

+++ philo/free.c
@@ -6,7 +6,7 @@
 /*   By: ilel-hla <ilel-hla@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/22 17:23:25 by ilel-hla          #+#    #+#             */
-/*   Updated: 2025/04/16 17:06:28 by ilel-hla         ###   ########.fr       */
+/*   Updated: 2025/06/18 22:21:16 by ilel-hla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void ft_destroy_mutexes(t_table *table)
 {
-	int i;
+	long i;
 
 	i = 0;
 	while (i < table->num_philos)
